+++ mypy-stubs/mistune/inline_parser.pyi
@@ -1,6 +1,3 @@
-__author__ = "Aleksandr Slepchenkov"
-__email__ = "Sl.aleksandr28@gmail.com"
-
 from typing import (
     Any,
     Dict,
@@ -14,57 +11,11 @@ from typing import (
     Type,
 )
 
-Tokens = List[Dict[str, Any]]
-# There are too much levels of optional unions of lists of text in cell and align 385 and 396 lines in mistune
-
-def escape(text: str, quote: bool = ..., smart_amp: bool = ...) -> str: ...
-
-class BlockGrammar:
-    def_links: Pattern[str]
-    def_footnotes: Pattern[str]
-    newline: Pattern[str]
-    block_code: Pattern[str]
-    fences: Pattern[str]
-    hrule: Pattern[str]
-    heading: Pattern[str]
-    lheading: Pattern[str]
-    block_quote: Pattern[str]
-    list_block: Pattern[str]
-    list_item: Pattern[str]
-    list_bullet: Pattern[str]
-    paragraph: Pattern[str]
-    block_html: Pattern[str]
-    table: Pattern[str]
-    nptable: Pattern[str]
-    text: Pattern[str]
-
-class BlockLexer:
-    grammar_class: Type[BlockGrammar]
-    default_rules: List[str]
-    list_rules: Tuple[str]
-    footnote_rules: Tuple[str]
-    tokens: Tokens
-    def_links: Dict[str, Dict[str, str]]
-    def_footnotes: Dict[str, int]
-    rules = ...  # type: BlockGrammar
-    def __init__(self, rules: Optional[BlockGrammar] = ..., **kwargs: Any) -> None: ...
-    def __call__(self, text: str, rules: Optional[Sequence[str]] = ...) -> Tokens: ...
-    def parse(self, text: str, rules: Optional[Sequence[str]] = ...) -> Tokens: ...
-    def parse_newline(self, m: Match[str]) -> None: ...
-    def parse_block_code(self, m: Match[str]) -> None: ...
-    def parse_fences(self, m: Match[str]) -> None: ...
-    def parse_heading(self, m: Match[str]) -> None: ...
-    def parse_lheading(self, m: Match[str]) -> None: ...
-    def parse_hrule(self, m: Match[str]) -> None: ...
-    def parse_list_block(self, m: Match[str]) -> None: ...
-    def parse_block_quote(self, m: Match[str]) -> None: ...
-    def parse_def_links(self, m: Match[str]) -> None: ...
-    def parse_def_footnotes(self, m: Match[str]) -> None: ...
-    def parse_table(self, m: Match[str]) -> None: ...
-    def parse_nptable(self, m: Match[str]) -> None: ...
-    def parse_block_html(self, m: Match[str]) -> None: ...
-    def parse_paragraph(self, m: Match[str]) -> None: ...
-    def parse_text(self, m: Match[str]) -> None: ...
+HTML_TAGNAME: str
+HTML_ATTRIBUTES: str
+ESCAPE_CHAR: Pattern[str]
+LINK_TEXT: str
+LINK_LABEL: str
 
 class InlineGrammar:
     escape: Pattern[str]
@@ -122,7 +73,7 @@ class InlineLexer:
     def output_strikethrough(self, m: Match[str]) -> str: ...
     def output_text(self, m: Match[str]) -> str: ...
 
-class Renderer:
+class InlineParser:
     options: Dict[str, str]
     def __init__(self, **kwargs: Any) -> None: ...
     def placeholder(self) -> str: ...
@@ -156,42 +107,3 @@ class Renderer:
     def footnote_ref(self, key: Any, index: int) -> str: ...
     def footnote_item(self, key: Any, text: str) -> str: ...
     def footnotes(self, text: Any) -> str: ...
-
-class Markdown:
-    renderer = ...  # type: Renderer
-    inline = ...  # type: InlineLexer
-    block = ...  # type: BlockLexer
-    footnotes = ...  # type: List[Dict[str, Any]]
-    tokens = ...  # type: Tokens
-    def __init__(
-        self,
-        renderer: Optional[Renderer] = ...,
-        inline: Optional[InlineLexer] = ...,
-        block: Optional[BlockLexer] = ...,
-        **kwargs: Any
-    ) -> None: ...
-    def __call__(self, text: str) -> str: ...
-    def render(self, text: str) -> str: ...
-    def parse(self, text: str) -> str: ...
-    token = ...  # type: Dict[str, Any]
-    def pop(self) -> Optional[Dict[str, Any]]: ...
-    def peek(self) -> Optional[Dict[str, Any]]: ...
-    def output(self, text: str, rules: Optional[Sequence[str]] = ...) -> str: ...
-    def tok(self) -> str: ...
-    def tok_text(self) -> str: ...
-    def output_newline(self) -> str: ...
-    def output_hrule(self) -> str: ...
-    def output_heading(self) -> str: ...
-    def output_code(self) -> str: ...
-    def output_table(self) -> str: ...
-    def output_block_quote(self) -> str: ...
-    def output_list(self) -> str: ...
-    def output_list_item(self) -> str: ...
-    def output_loose_item(self) -> str: ...
-    def output_footnote(self) -> str: ...
-    def output_close_html(self) -> str: ...
-    def output_open_html(self) -> str: ...
-    def output_paragraph(self) -> str: ...
-    def output_text(self) -> str: ...
-
-def markdown(text: str, escape: bool = ..., **kwargs: Any) -> str: ...
