+++ mypy-stubs/mistune/utils.pyi
@@ -1,0 +1,7 @@
+PUNCTUATION: str
+ESCAPE_TEXT: str
+
+def escape(text: str, quote: bool = True) -> str: ...
+def escape_html(s: str) -> str: ...
+def escape_url(s: str) -> str: ...
+def unikey(s: str) -> str: ...
